keyword use uint8_t

+++ pkg/ant.rmlui/src/css/PropertyKeyword.h
@@ -2,7 +2,7 @@
 
 namespace Rml {
 
-using PropertyKeyword = int;
+using PropertyKeyword = uint8_t;
 
 }
